pyproject.toml
[project]
name = "cosmos-curator"
version = "0.1.0"
description = "Cosmology video curation pipeline for dataset creation"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ray[data]>=2.40",
    "webdataset>=0.2",
    "yt-dlp>=2024.1",
    "ffmpeg-python>=0.2",
    "torch>=2.5",
    "open-clip-torch>=2.28",
    "faster-whisper>=1.1",
    "pydantic-settings>=2.6",
    "typer>=0.15",
    "numpy>=1.26",
    "imagehash>=4.3",
    "pillow>=10.0",
    "pyyaml>=6.0",
]

[project.scripts]
cosmos-curator = "run:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["cosmos_curator"]
